==> Method - 2/1. Alternate Scrapping.ipynb <==
import psaw
import praw
import time
import calendar
import pandas as pd

client_id = "zMvnVRM8MN_MRw"
client_secret = "-JIqik-pmfgku1BVB_2gcoxnMwM"
user_agent = "FlareReddit"
username = "psatreddit"
password = "pankaj123"

reddit = praw.Reddit(client_id = client_id, client_secret = client_secret, user_agent = user_agent, username = username, password = password)
api = psaw.PushshiftAPI(reddit)

epoch_time = calendar.timegm(time.gmtime())
gen = api.search_submissions(subreddit='india', before = epoch_time, limit=30000)
flairs = ["Politics", "Non-Political", "AskIndia", "Policy/Economy", "Business/Finance", "Science/Technology", "Scheduled", "Sports", "Food", "Photography", "CAA-NRC-NPR", "Coronavirus"]
print("Total No. Of Flairs:- ", len(flairs))

rid = []
title = []
url = []
score = []
comments_num = []
author = []
content = []
adultcontent = []
link_flare = []

for i in gen:
    post = vars(i)
    for flare in flairs:
        if post['link_flair_text'] == flare:
            rid.append(post['id'])
            title.append(post['title'])
            score.append(post['score'])
            url.append(post['permalink'])
            comments_num.append(post['num_comments'])
            author.append(post['author'])
            content.append(post['selftext'])
            adultcontent.append(post['over_18'])
            link_flare.append(post['link_flair_text'])

data = pd.DataFrame(rid, columns=["RID"])
data['Title'] = title
data['URL'] = url
data['Score'] = score
data['Comment_Score'] = comments_num
data['Author'] = author
data['Content'] = content
data['Adult'] = adultcontent
data['Flair'] = link_flare

data.to_csv("Reddit Data AS.csv")
