==> Method - 2/4. Alternate Modelling(Machine Learning).ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import SGDClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.pipeline import Pipeline
from sklearn.linear_model import LogisticRegression
from sklearn.svm import LinearSVC
from sklearn.neural_network import MLPClassifier
from xgboost import XGBClassifier
from sklearn.metrics import accuracy_score, classification_report
from sklearn.feature_extraction.text import CountVectorizer, TfidfTransformer

data = pd.read_csv('Cleaned Reddit Data AS.csv')
data.drop("Unnamed: 0", axis=1, inplace=True)
data.Content = data.Content.astype(str)
data.Title = data.Title.astype(str)
data.head()

# 1. Logistic Regression

def log_classifier(xtrain, xtest, ytrain, ytest):

    from sklearn.linear_model import LogisticRegression

    model_log = Pipeline([('vect', CountVectorizer()), 
                          ('tfidf', TfidfTransformer()), 
                          ('clf', LogisticRegression(n_jobs=1, C=1e5))
                         ])
    model_log.fit(xtrain, ytrain)

    y_pred = model_log.predict(xtest)

    print('accuracy %s' % accuracy_score(ytest, y_pred))
    print(classification_report(ytest, y_pred))

# 2. Multinomial Naive Bayes Classifier

def nb_classifier(xtrain, xtest, ytrain, ytest):
    model_nb = Pipeline([('vect', CountVectorizer()),
                         ('tfidf', TfidfTransformer()),
                         ('clf', MultinomialNB()),
                        ])
    model_nb.fit(xtrain, ytrain)

    y_pred = model_nb.predict(xtest)

    print('accuracy %s' % accuracy_score(ytest, y_pred))
    print(classification_report(ytest, y_pred))

# 3. Multi Layer Perceptron Classifier

def mlp_classifier(xtrain, xtest, ytrain, ytest):  
    model_mlp = Pipeline([('vect', CountVectorizer()),
                          ('tfidf', TfidfTransformer()),
                          ('clf', MLPClassifier(hidden_layer_sizes=(30,30,30))),
                         ])
    model_mlp.fit(xtrain, ytrain)

    y_pred = model_mlp.predict(xtest)

    print('accuracy %s' % accuracy_score(ytest, y_pred))
    print(classification_report(ytest, y_pred))

# 4. Random Forest Classifier

def rforest_classifier(xtrain, xtest, ytrain, ytest):
    model_rforest = Pipeline([('vect', CountVectorizer()),
                              ('tfidf', TfidfTransformer()),
                              ('clf', RandomForestClassifier(n_estimators = 1000, random_state = 42)),
                             ])
    model_rforest.fit(xtrain, ytrain)

    y_pred = model_rforest.predict(xtest)

    print('accuracy %s' % accuracy_score(ytest, y_pred))
    print(classification_report(ytest, y_pred))

# 5. Linear Support Vector Machine Classifier

def linear_svm_classifier(xtrain, xtest, ytrain, ytest):
    model_linearsvc = Pipeline([('vect', CountVectorizer()),
                                ('tfidf', TfidfTransformer()),
                                ('clf', LinearSVC()),
                               ])
    model_linearsvc.fit(xtrain, ytrain)

    y_pred = model_linearsvc.predict(xtest)

    print('accuracy %s' % accuracy_score(ytest, y_pred))
    print(classification_report(ytest, y_pred))

# 6. Stochastic Gradient Descent Classifier

def sgd_classifier(xtrain, xtest, ytrain, ytest):
    model_sgd = Pipeline([('vect', CountVectorizer()),
                          ('tfidf', TfidfTransformer()),
                          ('clf', SGDClassifier()),
                         ])
    model_sgd.fit(xtrain, ytrain)

    y_pred = model_sgd.predict(xtest)

    print('accuracy %s' % accuracy_score(ytest, y_pred))
    print(classification_report(ytest, y_pred))

# 7. XGBoost Classifier

def xgb_classifier(xtrain, xtest, ytrain, ytest):
    model_xgb = Pipeline([('vect', CountVectorizer()),
                          ('tfidf', TfidfTransformer()),
                          ('clf', XGBClassifier()),
                         ])
    model_xgb.fit(xtrain, ytrain)

    y_pred = model_xgb.predict(xtest)

    print('accuracy %s' % accuracy_score(ytest, y_pred))
    print(classification_report(ytest, y_pred))

def train_test(x,y):
 
    xtrain, xtest, ytrain, ytest = train_test_split(x, y, test_size=0.2)
   
    print("Results of Logistic Regression Classifier")
    log_classifier(xtrain, xtest, ytrain, ytest)
    print("Results of Multinomial Naive Bayes Classifier")
    nb_classifier(xtrain, xtest, ytrain, ytest)
    print("Results of Multi Layer Perceptron Classifier")
    mlp_classifier(xtrain, xtest, ytrain, ytest)
    print("Results of Random Forest Classifier")
    rforest_classifier(xtrain, xtest, ytrain, ytest)
    print("Results of Linear Support Vector Machine Classifier")
    linear_svm_classifier(xtrain, xtest, ytrain, ytest)
    print("Results of Stochastic Gradient Descent Classifier")
    sgd_classifier(xtrain, xtest, ytrain, ytest)
    print("Results of XGBoost Classifier")
    xgb_classifier(xtrain, xtest, ytrain, ytest)

flr = data.Flair
W = data.Title
X = data.URL
Y = data.Content
Z = W+X+Y

print("Flair Detection using Title as Feature")
train_test(W,flr)
print("Flair Detection using URL as Feature")
train_test(X,flr)
print("Flair Detection using Content as Features")
train_test(Y,flr)
print("Flair Detection using Combined Features")
train_test(Z,flr)
